% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/bindPosterior.R
\name{bindPosterior}
\alias{bindPosterior}
\title{Bind Posterior Distributions}
\usage{
bindPosterior(hydePost, relabel_factor = TRUE)
}
\arguments{
\item{hydePost}{An object of class \code{HydePosterior}}

\item{relabel_factor}{Logical.  If \code{TRUE}, factors that had been
  converted to integers for the JAGS code can be relabelled as factors
  for additional analysis in R.}
}
\description{
After determining the posterior distributions are satisfactory,
  it can be advantageous to bind the posterior distributions together in
  order to aggregate values and perform other manipulations and analyses.
}
\details{
For the purposes of this function, it is assumed that if the
  posterior distributions are satisfactory, the multiple chains in a run
  can be bound together.  Subsequently, the multiple runs are bound
  together.  Lastly, the factors are relabeled, if requested.
}
\author{
Jarrod Dalton and Benjamin Nutter
}

