% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/compileJagsModel.R
\name{compileJagsModel}
\alias{compileJagsModel}
\title{Compile Jags Model from a Hyde Network}
\usage{
compileJagsModel(network, data = NULL, ...)
}
\arguments{
\item{network}{An object of class \code{HydeNetwork}}

\item{data}{A list of data values to be observed in the nodes.  It is
passed to the \code{data} argument of \code{rjags::jags}.}

\item{...}{Additional arguments to be passed to \code{jags.model}}
}
\value{
Returns a \code{compiledHydeNetwork} object.  The \code{jags} element
  of this object is suitable to pass to \code{coda.samples}.  Otherwise,
  the primary function of the object is plotting the network with
  observed data shown.
}
\description{
Generates the JAGS code from the Hyde network and uses it to
  create an object representing a Bayesian graphical model.
}
\details{
\code{compileJagsModel} is a partial wrapper for
  \code{jags.model}. Running \code{compileJagsModel(network)} is
  equivalent to running \code{jgas.model(textConnection(writeNetworkModel(network)))}.
}
\examples{
mtcars2 <- transform(mtcars,
                     am=factor(am),
                     cyl=factor(cyl),
                     gear=factor(gear))
carNet <- HydeNetwork( ~ cyl +
                      disp | cyl +
                      hp | disp +
                      wt +
                      gear +
                      mpg | disp*hp*wt*gear,
                      data=mtcars2)

compiledCar <- compileJagsModel(carNet, n.chains=5)
s <- coda.samples(compiledCar$jags, c("cyl", "mpg"), n.iter=1000)
}
\author{
Benjamin Nutter
}
\seealso{
\code{jags.model}
}

