% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cpt.R
\name{cpt}
\alias{cpt}
\title{Compute a conditional probablity table for a factor given other factors}
\usage{
cpt(formula, data, wt)
}
\arguments{
\item{formula}{a formula with the following structure: \emph{response ~ var1 + var2 + ...}}

\item{data}{a data frame containing all the factors represented by the \code{formula}
parameter.}

\item{wt}{(optional) a numeric vector of observation weights.}
}
\description{
The function \code{cpt} operates on sets of factors. Specifically,
  it computes the conditional probability distribution of one of the factors
  given other factors, and stores the result in a multidimensional \code{array}.
}
\examples{
# a very imbalanced dice example

n <- 50000
data <- data.frame(
di1=as.factor(1:6 \%*\% rmultinom(n,1,prob=c(.4,.3,.15,.10,.03,.02))),
di2=as.factor(1:6 \%*\% rmultinom(n,1,prob=rev(c(.4,.3,.15,.10,.03,.02)))),
di3=as.factor(1:6 \%*\% rmultinom(n,1,prob=c(.15,.10,.02,.3,.4,.03)))
)
cpt <- cpt(di3 ~ di1 + di2, data)
plyr::aaply(cpt, c(1,2), sum)
}
\author{
Jarrod Dalton and Benjamin Nutter
}

