% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/factorRegex.R
\name{factorRegex}
\alias{factorRegex}
\title{Produce Regular Expressions for Extracting Factor Names and Levels}
\usage{
factorRegex(fit)
}
\arguments{
\item{fit}{a model object}
}
\description{
A utility function to produce a regular expression that can
  separate factor names and factor levels in the \code{broom::tidy()$term}
  output.
}
\author{
Benjamin Nutter
}

