% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HydeNet-package.R
\docType{package}
\name{HydeNet-package}
\alias{HydeNet-package}
\title{HydeNet: Hybrid Bayesian Networks Using R and JAGS}
\description{
Facilities for easy implementation of hybrid Bayesian networks using R. Bayesian networks are directed acyclic graphs representing joint probability distributions, where each node represents a random variable and each edge represents conditionality. The full joint distribution is therefore factorized as a product of conditional densities, where each node is assumed to be independent of its non-descendents given information on its parent nodes. Since exact, closed-form algorithms are computationally burdensome for inference within hybrid networks that contain a combination of continuous and discrete nodes, particle-based approximation techniques like Markov Chain Monte Carlo are popular. We provide a user-friendly interface to constructing these networks and running inference using the 'rjags' package. Econometric analyses (maximum expected utility under competing policies, value of information) involving decision and utility nodes are also supported.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/nutterb/HydeNet,}
  \item Report bugs at \url{https://github.com/nutterb/HydeNet/issues}
}

}
\keyword{internal}
