% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HydeNet-package.R
\name{HydeNet}
\alias{HydeNet}
\alias{Hyde}
\title{Hydbrid Decision Networks
burdensome for inference within hybrid networks that contain a combination 
each edge represents conditionality. The full joint distribution is 
Econometric analyses (maximum expected utility under competing policies, 
interface to constructing these networks and running inference using rjags. 
is assumed to be independent of its non-descendants given information on its 
like Markov Chain Monte Carlo are popular. We provide a user-friendly 
of continuous and discrete nodes, particle-based approximation techniques 
parent nodes. Since exact, closed-form algorithms are computationally 
probability distributions, where each node represents a random variable and 
supported.
therefore factorized as a product of conditional densities, where each node 
using R. Bayesian networks are directed acyclic graphs representing joint 
value of information) involving decision and utility nodes are also}
\description{
Facilities for easy implementation of hybrid Bayesian networks
}
