% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summary.HydeNetwork.R
\name{HydeNetSummaries}
\alias{HydeNetSummaries}
\alias{summary.HydeNetwork}
\title{HydeNet Summary Objects}
\usage{
\method{summary}{HydeNetwork}(object, ...)
}
\arguments{
\item{object}{A \code{HydeNet} object to be summarized}

\item{...}{Additional arguments.}
}
\description{
Summaries of \code{HydeNetwork}, compiled network, and
  compiled decision network objects.
}
\author{
Jarrod Dalton and Benjamin Nutter
}

