% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/HydePosterior.R
\name{HydePosterior}
\alias{HydePosterior}
\title{Posterior Distributions of a Decision Network}
\usage{
HydePosterior(cHN, variable.names, n.iter, thin = 1, ...,
  monitor_observed = TRUE)
}
\arguments{
\item{cHN}{A \code{compiledHydeNetwork} object as returned by
\code{compileJagsNetwork}.}

\item{variable.names}{a character vector giving the names of variables to be monitored.}

\item{n.iter}{number of iterations to monitor.}

\item{thin}{thinning interval for monitors.}

\item{...}{options arguments that are passed to the update method for
  jags model objects.}
}
\value{
A list of class \code{HydePosterior} with elements \code{codas}
  (the MCMC matrices from \code{coda.samples}), \code{observed} (the values
  of the variables that were observed), \code{dag} (the dag object for
  convenience in displaying the network), and \code{factorRef} (giving the
  mappings of factor levels to factor variables).

  The only rationale for giving this object its own class was because it
  produces an enormous amount of material to be printed.  A distinct
  \code{print} method has been written for this object.
}
\description{
The posterior distributions of the decision network can be
  evaluated to determine the probabilistic outcomes based on the decision
  inputs in the model as well as subject specific factors.
}
\details{
This is essentially a wrapper around \code{coda.samples} that
  returns in a list the output for each run of \code{coda.samples} over
  the rows of the policy/decision matrix given in the \code{data} argument
  of \code{compileJagsNetwork}.
}
\author{
Jarrod Dalton and Benjamin Nutter
}

