% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/HydeUtilities.R
\name{HydeUtilities}
\alias{HydeUtilities}
\alias{makeJagsReady}
\alias{matchLevelNumber}
\alias{termName}
\alias{validateParameters}
\title{Hyde Network Utility Functions}
\usage{
termName(term, reg)

makeJagsReady(mdl, regex)

matchLevelNumber(t, Ref)

validateParameters(params, dist)
}
\arguments{
\item{term}{Usually the \code{term} column from the output of
\code{broom::tidy()}}

\item{reg}{A regular expression, usually provided by \code{factorRegex}}

\item{mdl}{Output from \code{broom::tidy()}}

\item{regex}{A regular expression, usually returned by \code{factorRegex}}

\item{t}{Usually the \code{term_name} column generated within
\code{makeJagsReady}}

\item{Ref}{usually the levels reference data frame generated within
\code{makeJagsReady}}

\item{params}{The list of parameters given in the \code{...} argument of
\code{setNode}}

\item{dist}{The JAGS distribution function name.  Appropriate names are
in the \code{FnName} column of the \code{jagsDists} data object.}
}
\description{
The functions described below are unexported functions that
  are used internally by \code{HydeNet} to prepare modify network objects
  and prepare JAGS code.
}
\details{
\code{termName}: In most model objects, factors in the model are
  represented as [variable][level].  This utility function pulls out the
  [variable] component.  This is typically called from within
  \code{makeJagsReady}.

  \code{makeJagsReady}: Manages the presence of factors in interactions and
  makes sure the proper numeric factor value is given to the JAGS code.
  This is called from within a \code{writeJagsFormula} call.

  \code{matchLevelNumber}: Assigns the correct numeric value of a level to
  a factor variable in a model.  This is called from within
  \code{makeJagsRead}.

  \code{validateParameters}: Users may pass parameters to the JAGS code using the
  \code{setNode} function.  If a faulty parameter is given, such as
  \code{lambda = -10} for a poisson distribution (\code{lambda} must be
  positive in a Poisson distribution), the error returned by JAGS may not
  clear enough to diagnose a problem that presented several steps earlier
  in the code.  \code{validateParamters} allows the user to receive instant
  feedback on the appropriateness of the code.

  Logical expressions for comparisons are stored in the \code{jagsDists}
  data object (\code{data(jagsDists, package='Hyde')}).  This is a utility
  function used only within \code{setNode} and is not visible to the user.
}
\author{
Jarrod Dalton and Benjamin Nutter
}

