% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{inputCPTExample}
\alias{inputCPTExample}
\title{Example Conditional Probability Table Resulting from the \code{inputCPT} function.}
\format{\preformatted{ cpt [1:2, 1:2] 0.77 0.15 0.23 0.85
 - attr(*, "dimnames")=List of 2
  ..$ disease: chr [1:2] "Healthy" "Diseased"
  ..$ test   : chr [1:2] "---" "+++"
 - attr(*, "model")='data.frame':	4 obs. of  4 variables:
  ..$ disease: Factor w/ 2 levels "Healthy","Diseased": 1 2 1 2
  ..$ test   : Factor w/ 2 levels "---","+++": 2 2 1 1
  ..$ wt     : num [1:4] 0.23 0.85 0.77 0.15
  ..$ wt     : num [1:4] 0.23 0.85 0.77 0.15
 - attr(*, "class")= chr [1:2] "cpt" "array"
}}
\source{
No Source.  It's really just made up.
}
\usage{
inputCPTExample
}
\description{
This is an example of the output generated by the \code{inputCPT} function as
illustrated in the article being submitted to JSS.  It is saved as an object
named \code{h} in the article.
}
\keyword{datasets}

