% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.HydeNetwork.R
\name{mergeDefaultPlotOpts}
\alias{mergeDefaultPlotOpts}
\title{rdname plot.HydeNetwork}
\usage{
mergeDefaultPlotOpts(network, node_df)
}
\arguments{
\item{network}{a \code{HydeNetwork} object}

\item{node_df}{A data frame of node attributes.}
}
\description{
rdname plot.HydeNetwork
}

