% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot.compiledHydeNetwork}
\alias{plot.compiledHydeNetwork}
\alias{plotcompiledHydeNetwork}
\title{Plot a Compiled Hyde Network}
\usage{
\method{plot}{compiledHydeNetwork}(x, shape_observed = "rect",
  color_observed = "#8C6BB1", nodeAttrs = list(), ...)
}
\arguments{
\item{x}{An object of class \code{compiledHydeNetwork}}

\item{shape_observed}{The default shape for observed nodes}

\item{color_observed}{The default color for observed nodes}

\item{nodeAttrs}{A list of node attributes to be passed to
\code{graph::plot}.  See "Notes" for more details.}

\item{...}{Other arguments to be passed to \code{graph::plot}.}
}
\description{
Compiled Hyde Networks are network objects created by
  \code{compileJagsModel}.  The primary differentiation is that a
  compiled Network may have observed nodes.  In order to simplify the
  work of showing differences in compiled networks (and the observed nodes
  within them), this plotting method will change the color and shape of
  observed nodes by default.
}
\details{
Default colors and shapes are used for observed nodes (nodes for which at
  least one value was given in the \code{compileJagsModel} function's
  \code{data} argument) unless overridden by the user.  To suppress the
  defaults, it is easiest to use \code{shape_observed = 'circle'} and
  \code{color_observed = 'white'}.
}
\note{
For customizations, see the \code{Rgraphviz} documentation for
  \code{GraphvizAttributes}.  Most notably, read the sections about
  General Node Attributes to manipulate shapes and positions of the
  nodes and their labels.  These are manipulated through the \code{nodeAttrs}
  arguments (passed through \code{...}).

  General Graph Attributes are passed through \code{attrs} and Edge
  Attributes are passed through \code{edgeAttrs}.
}
\examples{
carNet <- HydeNetwork( ~ cyl +
                      disp | cyl +
                      hp | disp +
                      wt +
                      gear +
                      mpg | disp*hp*wt*gear,
                      data=mtcars)
plot(carNet)

compiledCar <- compileJagsModel(carNet, data=list(gear=4))
plot(compiledCar)
}
\author{
Jarrod Dalton and Benjamin Nutter
}

