% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.cpt.R
\name{print.cpt}
\alias{print.cpt}
\title{Print Method for CPT objects}
\usage{
\method{print}{cpt}(x, ...)
}
\arguments{
\item{x}{Object of class \code{cpt}}

\item{...}{Additional arguments to be passed to other methods.}
}
\description{
Just a wrapper to strip the attributes off, change the class, and print the array.
}
\author{
Jarrod Dalton and Benjamin Nutter
}

