% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/print.HydeNetwork.R
\name{print.HydeNetwork}
\alias{print.HydeNetwork}
\title{Print a HydeNetwork}
\usage{
\method{print}{HydeNetwork}(x, ...)
}
\arguments{
\item{x}{a \code{HydeNetwork object}}

\item{...}{additional arguments to be passed to print methods.  Currently
  none in use.}
}
\description{
Print a HydeNetwork
}
\details{
Prints a HydeNetwork object with a brief summary of each node.
}
\examples{
carNet <- HydeNetwork( ~ cyl +
                      disp | cyl +
                      hp | disp +
                      wt +
                      gear +
                      mpg | disp*hp*wt*gear,
                      data=mtcars)
carNet

carNet <- setNode(carNet, mpg,
                  nodeType='dnorm', mu=fromFormula(), tau=1/2.65,
                  nodeFormula = mpg ~ disp + hp + wt + factor(gear),
                  nodeFitter='lm')
carNet
}
\author{
Jarrod Dalton and Benjamin Nutter
}

