% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/print.HydePosterior.R
\name{print.HydePosterior}
\alias{print.HydePosterior}
\title{Print a Hyde Posterior Distribution Object}
\usage{
\method{print}{HydePosterior}(x, ...)
}
\arguments{
\item{x}{a \code{HydePosterior} object}

\item{...}{additional arguments to be passed to print methods.  Currently
  none in use.}
}
\description{
Print a Hyde Posterior Distribution Object
}
\details{
Prints a brief description of a HydePosterior object.
}
\author{
Jarrod Dalton and Benjamin Nutter
}

