% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Resolution.cpt}
\alias{Resolution.cpt}
\title{Conditional Probability Table for resolution of side effects as a function drugs and emesis.}
\format{An object of class \code{cpt} (inherits from \code{array}) of dimension 2 x 2 x 2 x 2 x 3.}
\usage{
Resolution.cpt
}
\description{
This is a conditional probability table used in the emesis example of the JSS article.
}
\keyword{datasets}
