% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SE.cpt}
\alias{SE.cpt}
\title{Conditional Probability Table for side effects as a function of emesis and drug.}
\format{\preformatted{ cpt [1:2, 1:2, 1:2] 0.95 0.92 0.85 0.73 0.05 0.08 0.15 0.27
 - attr(*, "dimnames")=List of 3
  ..$ Emesis: chr [1:2] "No" "Yes"
  ..$ Drug  : chr [1:2] "Metoclopramide" "Ondansetron"
  ..$ SE    : chr [1:2] "No" "Yes"
 - attr(*, "model")='data.frame':	8 obs. of  5 variables:
  ..$ Emesis: Factor w/ 2 levels "No","Yes": 1 2 1 2 1 1 2 2
  ..$ Drug  : Factor w/ 2 levels "Metoclopramide",..: 1 1 2 2 1 2 1 2
  ..$ SE    : Factor w/ 2 levels "No","Yes": 2 2 2 2 1 1 1 1
  ..$ wt    : num [1:8] 0.05 0.08 0.15 0.27 0.95 0.85 0.92 0.73
  ..$ wt    : num [1:8] 0.05 0.08 0.15 0.27 0.95 0.85 0.92 0.73
 - attr(*, "class")= chr [1:2] "cpt" "array"
}}
\usage{
SE.cpt
}
\description{
This is a conditional probability table used in the emesis example of the JSS article.
}
\keyword{datasets}

