% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fromData}
\alias{fromData}
\alias{fromFormula}
\alias{setNode}
\title{Set Node Relationships}
\usage{
fromData()

fromFormula()

setNode(network, node, nodeType = "dnorm", params, nodeFitter, nodeFormula,
  fromData = !is.null(network$data), ...)
}
\arguments{
\item{network}{A \code{HydeNetwork}.}

\item{node}{A node within \code{network}.  This does not have to be quoted.}

\item{nodeType}{a valid distribution function from JAGS.  See the data set
in \code{data(jagsDists)} for a complete list.}

\item{params}{to be passed to the JAGS distribution function.  Each parameter
in the distribution function must be a named element in the vector.  For
example, the parameters to pass to \code{dnorm} would be \code{c(mu='', tau='')}.
The required parameters can be looked up using the
\code{expectedParameters} function.  If parameters are to be estimated
from the data, the functions \code{fromData} and \code{fromFormula} may
be used as placeholders.}

\item{nodeFitter}{the fitting function, such as \code{lm} or \code{glm}.  This
will probably only be needed when \code{fromData = TRUE}.}

\item{nodeFormula}{A formula object specifying the relationship between a
node and its parents.  It must use as a term every parent of \code{node}.}

\item{fromData}{Logical.  Determines if a node's relationship is calculated
from the data object in \code{network}.  Defaults to \code{TRUE} whenever
\code{network} has a data object.}

\item{...}{Additional arguments to be passed to \code{fitter}.  This does not
  yet have any effect as I haven't yet decided out where to store this and
  how to implement the fitting.}
}
\description{
Set Node Relationships
}
\details{
The relationship between a node and its parents must be defined
  before the appropriate JAGS model statement can be constructed.
  \code{setNode} is the utility by which a user can define the distribution
  of the node and its relationship to its parents (usually through a model
  of some sort).
}
\examples{
carNet <- HydeNetwork( ~ cyl +
                      disp | cyl +
                      hp | disp +
                      wt +
                      gear +
                      mpg | disp*hp*wt*gear,
                      data=mtcars)

carNet$nodeFormula$mpg
carNet$nodeFitter$mpg
carNet$nodeType$mpg

carNet <- setNode(gm, mpg, nodeType='dnorm', params=c(tau=1/2.65),
                  nodeFormula = mpg ~ disp + hp + wt + factor(gear),
                  nodeFitter='lm')

carNet$nodeFormula$mpg
carNet$nodeFitter$mpg
carNet$nodeType$mpg
}
\author{
Jarrod Dalton and Benjamin Nutter
}

