% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/setNodeModels.R
\name{setNodeModels}
\alias{setNodeModels}
\title{Set Node Properties Using Model Objects}
\usage{
setNodeModels(network, ...)
}
\arguments{
\item{network}{A \code{HydeNetwork} object}

\item{...}{Model objects to be incorporated into \code{network}}
}
\description{
Set node properties using pre-defined model objects.  Model
  objects may be imported from other programs, but need to be valid
  model objects with the additional restriction that the responses and
  independent variables must be named nodes in the network.
}
\author{
Benjamin Nutter
}

