% Generated by roxygen2 (4.0.2): do not edit by hand
\name{validateParameters}
\alias{validateParameters}
\title{Validate JAGS Distributional Parameters}
\usage{
validateParameters(params, dist)
}
\arguments{
\item{params}{The list of parameters given in the \code{...} argument of
\code{setNode}}

\item{dist}{The JAGS distribution function name.  Appropriate names are
  in the \code{FnName} column of the \code{jagsDists} data object.}
}
\description{
Users may pass parameters to the JAGS code using the
  \code{setNode} function.  If a faulty parameter is given, such as
  \code{lambda = -10} for a poisson distribution (\code{lambda} must be
  positive in a Poisson distribution), the error returned by JAGS may not
  clear enough to diagnose a problem that presented several steps earlier
  in the code.  \code{validateParamters} allows the user to receive instant
  feedback on the appropriateness of the code.
}
\details{
Logical expressions for comparisons are stored in the \code{jagsDists}
  data object (\code{data(jagsDists, package='Hyde')}).  This is a utility
  function used only within \code{setNode} and is not visible to the user.
}
\author{
Jarrod Dalton and Benjamin Nutter
}

