% Generated by roxygen2 (4.0.2): do not edit by hand
\name{writeJagsFormula.glm}
\alias{writeJagsFormula}
\alias{writeJagsFormula.glm}
\alias{writeJagsFormula.lm}
\alias{writeJagsFormula.multinom}
\alias{writeJagsFormula.xtabs}
\title{Write the JAGS Formula for a Hyde Node}
\usage{
\method{writeJagsFormula}{glm}(fit, ...)

\method{writeJagsFormula}{lm}(fit, ...)

\method{writeJagsFormula}{multinom}(fit, ...)

writeJagsFormula(fit, ...)

\method{writeJagsFormula}{xtabs}(fit, ...)
}
\arguments{
\item{fit}{a model object}

\item{...}{Additional arguments to be passed to other methods}
}
\description{
Based on the information provided about the node,
  an appropriate JAGS model is written in text.  This is combined with
  the other node models to generate the complete network.
}
\details{
Methods for different model objects can be written so that
  this function can be extended as desired.

  In the \code{writeJagsFormula.glm} method, appropriate transformations
  exist for the following combinations:
  \enumerate{
    \item family = binomial; link = logit
    \item family = poisson; link = log
    \item family = gaussian; link = identity (calls \code{writeJagsFormula.lm})
  }
}
\author{
Jarrod Dalton and Benjamin Nutter
}

