% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{inputCPTExample}
\alias{inputCPTExample}
\title{Example Conditional Probability Table Resulting from the \code{inputCPT} function.}
\format{An object of class \code{cpt} (inherits from \code{array}) with 2 rows and 2 columns.}
\source{
No Source.  It's really just made up.
}
\usage{
inputCPTExample
}
\description{
This is an example of the output generated by the \code{inputCPT} function as 
illustrated in the article being submitted to JSS.  It is saved as an object 
named \code{h} in the article.
}
\keyword{datasets}

