% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{jagsDists}
\alias{jagsDists}
\title{JAGS Probability Distributions.}
\format{A data frame with 30 rows and 6 variables:
\describe{
  \item{DistName}{Distribution Name}
  \item{FnName}{Function Name}
  \item{xLow}{Minimum value for x, the random variable}
  \item{xHigh}{Maximum value for x, the random variable}
  \item{Parameters}{Names of the parameters}
  \item{paramLimit}{Limits on the parameter}
}}
\source{
\url{http://people.math.aau.dk/~kkb/Undervisning/Bayes14/sorenh/docs/jags_user_manual.pdf}
}
\usage{
jagsDists
}
\description{
A dataset listing the JAGS probability distributions and their parameters
}
\keyword{datasets}

