% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{jagsFunctions}
\alias{jagsFunctions}
\title{JAGS Functions Compatible with R.}
\format{A data frame with 30 rows and 3 variables:
\describe{
  \item{jags_function}{JAGS function name}
  \item{r_function}{R function Name}
  \item{r_package}{R package where the function is found.}
}}
\source{
\url{http://people.math.aau.dk/~kkb/Undervisning/Bayes14/sorenh/docs/jags_user_manual.pdf}
}
\usage{
jagsFunctions
}
\description{
A dataset listing the JAGS functions and their R equivalents.
}
\keyword{datasets}

