% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/modelToNode.R
\name{modelToNode}
\alias{modelToNode}
\alias{modelToNode.cpt}
\alias{modelToNode.default}
\alias{modelToNode.glm}
\alias{modelToNode.lm}
\alias{modelToNode.multinom}
\alias{modelToNode.survreg}
\alias{modelToNode.xtabs}
\title{Convert a Model Object to a Network Node}
\usage{
modelToNode(model, nodes, ...)

\method{modelToNode}{default}(model, nodes, ...)

\method{modelToNode}{cpt}(model, nodes, ...)

\method{modelToNode}{glm}(model, nodes, ...)

\method{modelToNode}{lm}(model, nodes, ...)

\method{modelToNode}{multinom}(model, nodes, ...)

\method{modelToNode}{survreg}(model, nodes, ...)

\method{modelToNode}{xtabs}(model, nodes, ...)
}
\arguments{
\item{model}{A model object}

\item{nodes}{A vector of node names, usually as \code{network$nodes}}

\item{...}{Additional arguments to be passed to other functions.  Currently ignored.}
}
\description{
In cases where model objects may already be fit and established,
  they can be used to generate a network without having to refit the
  models or specify the distributions in \code{setNode}.
}

