% Generated by roxygen2 (4.0.2): do not edit by hand
\name{modelToNode.default}
\alias{modelToNode}
\alias{modelToNode.default}
\alias{modelToNode.lm}
\alias{modelToNode.multinom}
\alias{modelToNode.xtabs}
\title{Convert a Model Object to a Network Node}
\usage{
\method{modelToNode}{default}(model, ...)

\method{modelToNode}{lm}(model, ...)

\method{modelToNode}{lm}(model, ...)

\method{modelToNode}{multinom}(model, ...)

modelToNode(model, ...)

\method{modelToNode}{xtabs}(model, ...)
}
\arguments{
\item{model}{A model object}

\item{...}{Additional arguments to be passed to other functions.  Currently ignored.}
}
\description{
In cases where model objects may already be fit and established,
  they can be used to generate a network without having to refit the
  models or specify the distributions in \code{setNode}.
}

