% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{PE}
\alias{PE}
\title{PE Data Set (Better Title)}
\format{A data frame with 10000 rows and 7 variables:
\describe{
  \item{wells}{}
  \item{pregnant}{}
  \item{pe}{}
  \item{angio}{}
  \item{d.dimer}{}
  \item{treat}{}
  \item{death}{}
}}
\source{
WHAT IS THE SOURCE?
}
\usage{
PE
}
\description{
Description of Data Set
}
\keyword{datasets}

