% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.HydeNetwork.R
\name{plot.HydeNetwork}
\alias{HydePlotOptions}
\alias{customNode}
\alias{mapEdges}
\alias{mergeCustomEdges}
\alias{mergeCustomNodes}
\alias{plot.HydeNetwork}
\alias{plotHydeNetwork}
\title{Plotting Utilities Probabilistic Graphical Network}
\source{
\url{http://rich-iannone.github.io/DiagrammeR/graphviz.html}\cr
  See especially the section on Attributes

  \url{http://graphviz.org/}\cr
  \url{http://graphviz.org/content/attrs}
}
\usage{
\method{plot}{HydeNetwork}(x, customNodes = NULL, customEdges = NULL, ...,
  useHydeDefaults = TRUE)

mergeCustomNodes(node_df, customNodes)

mapEdges(n, p)

mergeCustomEdges(edge_df, customEdges)

customNode(node_id, ...)

HydePlotOptions(variable = NULL, determ = NULL, decision = NULL,
  utility = NULL, restorePackageDefaults = FALSE)
}
\arguments{
\item{x}{an object of class \code{HydeNetwork}}

\item{customNodes}{a data frame giving additional specifications for nodes.
The customizations provided here will override the default settings.}

\item{customEdges}{a data frame giving custom settings for edges (arrows)
between nodes.}

\item{...}{for the \code{plot} method, additional arguments to be passed to
\code{DiagrammeR::graphviz_graph}.  For \code{customNode},
named node attributes to assign to a node's plotting characteristics.}

\item{useHydeDefaults}{A logical value indicating if the default plot
  parameters in \code{options("Hyde_plotOptions")} should be applied
  to the plot.}

\item{node_df}{A data frame of node attributes}

\item{n}{node names from a network object}

\item{p}{the list of parents from a network object}

\item{edge_df}{The default edge attribute data frame}

\item{node_id}{The name of a node in a \code{HydeNetwork} object.
May be quoted or unquoted.}

\item{variable,determ,decision,utility}{Named lists of attributes to use as
defaults node attribute settings for each variable type.}

\item{restorePackageDefaults}{A logical value.  When TRUE, the original
package defaults are restored.}
}
\description{
Plotting Utilities Probabilistic Graphical Network
}
\details{
Generate and customize plots of a \code{HydeNetwork}
  class network. \code{HydeNet} provides some initial defaults for standard
  variable nodes, deterministic nodes, decision nodes, and utility nodes.
  Since these nodes are assumed to be of inherent difference and interest,
  the options are defined in a way to make these easier to identify in
  a plot.  The default options may be altered by the user to their liking
  by invoking \code{HydePlotOptions}.  Node attributes are more fully
  explained in the documentation for the \code{DiagrammeR} package.
  Individual nodes may be define with \code{customNode}.
}
\examples{
\dontrun{
#* Plots may open in a browser.
data(BlackJack, package="HydeNet")
plot(BlackJack)

HydePlotOptions(variable=list(shape = "rect", fillcolor = "green"),
                determ = list(shape = "rect", fillcolor = "black",
                              fontcolor = "white", linecolor = "white"),
                decision = list(shape = "ellipse", fillcolor = "yellow",
                                linecolor = "red"),
                utility = list(shape = "circle", fillcolor = "red",
                               fontcolor = "white"))
plot(BlackJack)

HydePlotOptions(restorePackageDefaults = TRUE)

plot(BlackJack,
     customNodes = customNode(fillcolor = "purple", shape = "circle",
                              fontcolor = "white", height = "2",
                              style="filled"))
plot(BlackJack,
  customNodes =
    dplyr::bind_rows(
      customNode(pointsAfterCard3,
                 shape = "circle",
                 style = "radial",
                 fillcolor = "purple:yellow",
                 fontcolor = "black",
                 height = "2"),
      customNode(playerFinalPoints,
                 shape = "circle",
                 style = "wedged",
                 height = "3",
                 fillcolor = c("red:orange:yellow:green:blue:purple"))))
}
}
\author{
Jarrod Dalton and Benjamin Nutter
}

