% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Resolution.cpt}
\alias{Resolution.cpt}
\title{Conditional Probability Table for resolution of side effects as a function drugs and emesis.}
\format{\preformatted{ cpt [1:2, 1:2, 1:2, 1:2, 1:3] 1 0 1 0 1 0 1 0 1 0 ...
 - attr(*, "dimnames")=List of 5
  ..$ SE        : chr [1:2] "No" "Yes"
  ..$ TreatSE   : chr [1:2] "No" "Yes"
  ..$ Emesis    : chr [1:2] "No" "Yes"
  ..$ Drug      : chr [1:2] "Metoclopramide" "Ondansetron"
  ..$ Resolution: chr [1:3] "NoSE" "Resolved" "Unresolved"
 - attr(*, "model")='data.frame':	48 obs. of  7 variables:
  ..$ SE        : Factor w/ 2 levels "No","Yes": 1 2 1 2 1 2 1 2 1 2 ...
  ..$ TreatSE   : Factor w/ 2 levels "No","Yes": 1 1 2 2 1 1 2 2 1 1 ...
  ..$ Emesis    : Factor w/ 2 levels "No","Yes": 1 1 1 1 2 2 2 2 1 1 ...
  ..$ Drug      : Factor w/ 2 levels "Metoclopramide",..: 1 1 1 1 1 1 1 1 2 2 ...
  ..$ Resolution: Factor w/ 3 levels "NoSE","Resolved",..: 1 1 1 1 1 1 1 1 1 1 ...
  ..$ wt        : num [1:48] 1 0 1 0 1 0 1 0 1 0 ...
  ..$ wt        : num [1:48] 1 0 1 0 1 0 1 0 1 0 ...
 - attr(*, "class")= chr [1:2] "cpt" "array"
}}
\usage{
Resolution.cpt
}
\description{
This is a conditional probability table used in the emesis example of the JSS article.
}
\keyword{datasets}

