% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SE.cpt}
\alias{SE.cpt}
\title{Conditional Probability Table for side effects as a function of emesis and drug.}
\format{
An object of class \code{cpt} (inherits from \code{array}) of dimension 2 x 2 x 2.
}
\usage{
SE.cpt
}
\description{
This is a conditional probability table used in the emesis example of the JSS article.
}
\keyword{datasets}
