% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/setDecisionNodes.R
\name{setDecisionNodes}
\alias{setDecisionNodes}
\alias{setUtilityNodes}
\title{Classify Multiple Nodes as Decision or Utility Nodes}
\usage{
setDecisionNodes(network, ...)

setUtilityNodes(network, ...)
}
\arguments{
\item{network}{A Hyde Network object}

\item{...}{Networks to be classified as decision nodes.  These may be quoted
  or unquoted.}
}
\description{
Depending on how your Hyde Network was built, you may not have
  had the opportunity to declare which nodes are decision or utlity nodes.
  For instance,
  when passing a list of models, there is no way to indicate in the model
  object that the node should be considered a decision node.  As a matter
  of convenience, these function will set any nodes indicated to decision
  or utility nodes.  It will make no other modifications to a node's definition.
}
\author{
Jarrod Dalton and Benjamin Nutter
}

