% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/update.HydeNetwork.R
\name{update.HydeNetwork}
\alias{update.HydeNetwork}
\title{Update Probabilistic Graphical Network}
\usage{
\method{update}{HydeNetwork}(object, formula, ...)
}
\arguments{
\item{object}{A \code{HydeNetwork} object}

\item{formula}{A formula statement indicating the changes to the network.}

\item{...}{Additional arguments to be passed to other methods.  Current,
  none are used.}
}
\description{
Add or remove nodes or add parents within a \code{HydeNetwork}
  model.
}
\details{
Adding or removing nodes is fairly straightforward if you are
  removing a complete node (along with its parents).  If you just wish to
  remove a parent, this doesn't work well yet.  I will have to work out
  a solution to remove parent relationships individually.  I had hoped
  there would be an \code{update.dag} method, but no luck.
}
\examples{
mtcars2 <- transform(mtcars,
                     am=factor(am),
                     cyl=factor(cyl),
                     gear=factor(gear))
carNet <- HydeNetwork( ~ cyl +
                      disp | cyl +
                      hp | disp +
                      wt +
                      gear +
                      mpg | disp*hp*wt*gear,
                      data=mtcars2)

plot(carNet)

carNet2 <- update(carNet, ~ . + am|cyl + mpg|am)
plot(carNet2)
}
\author{
Jarrod Dalton and Benjamin Nutter
}

